백엔드 Request 등 리팩토링

+++ src/main/java/com/itschool/tableq/controller/api/ReservationApiController.java
@@ -49,7 +49,7 @@ public class ReservationApiController extends
 
     @Operation(summary = "예약의 현재 순번 조회", description = "Restaurant ID 와 Reservation ID로 현재 순번 조회")
     @GetMapping("/user/queue-left/{reservationId}")
-    public Header<Integer> readMyQueue(@PathVariable(name = "restaurantId") Long restaurantId,
+    public Header<Long> readMyQueue(@PathVariable(name = "restaurantId") Long restaurantId,
                                      @PathVariable(name = "reservationId") Long reservationId){
         log.info("{}","{}","{}","read: ",restaurantId, reservationId);
         try {
